WIP Test `SetStepMode`
This is my in-progress code for the `SetStepMode` test. I rebased this
on top of the latest test stand code and made sure it compiled, but I
didn't test it again. Last time I tried to get this working, there were
all kinds of problems (see #26).

I plan to pick this up again at a later time.