Add traits for enabling driver capabilities
More traits mean more complexity, but I think this is justified.

First, this is going to solve one immediate problem: When enabling
control over step mode or direction, you would want to set the initial
step mode or direction immediately, to start with a defined value.

For step mode, this is currently done, but it would no longer work once
the duplicated code in the `set_step_mode` implementation is moved to
`Driver`, since the driver itself would no longer be able to call
`self.set_step_mode` in its driver-specific enable method. With these
traits, `Driver` will be able to take full control here.

Second, this allows `Driver` to abstract over more aspects of the
driver. Once the potential provided by these traits has been fully
realized, users will no longer have to call anything on the
driver-specific structs. The most they have to care about, is which
driver-specific resources they have to pass to the `Driver` methods.

Third, this is going to provide a lot of flexibility. This flexibility
will be nice to have in some cases, when users want very specific
control over which driver capabilities they enable, and which they
control via other means.

But this will also make help make some future capabilities possible that
are not really possible right now. Some drivers, like the STSPIN220,
have pins with overlapping meaning. This is fine right now, as
everything is an `OutputPin` anyway. But once we support PWM, a user
might have to enable mode control, set the mode, then disable it again
(which can be made possible by the complementing "disable" traits), so
it can re-configure the `OutputPin` as a PWM pin and enable step control
using this PWM pin.