Make `Driver::step` non-blocking
The convenience of the current API is preserved with the
`StepFuture::wait` function. Otherwise, this commit introduces a
significant amount of complexity.

I think that's worth it, given the new capability. Especially as I don't
know how to do any better right now, and since this is bound to get
_much_ better (both for the user and in the implementation), once the
embedded ecosystem grows sufficient support for async/await.