Depend on concrete timer in test stand
Updates the test stand code to depend on a concrete timer instead of the
trait from embedded-hal. The previous approach wasn't that great, as the
timer used is known anyway, and the methods weren't totally generic to
begin with.

This has the side effect of fixing the build with the latest version of
LPC8xx HAL's master branch.