Change main branch name to `main`
There's a political argument for changing the name of the main branch
from `master` to `main`. I don't know what to think about that, so I'm
going to ignore it for my following arguments.

I think "main" is clearly the better name for the main/master branch:

- It is shorter and has half the number of syllables.
- It is a much clearer name. At least to me as a non-native speaker, the
  concept of a master record, which is the basis of the name "master",
  as far as I understand, is much more obscure than "main", which is a
  basic and pervasive word.

Given that, I see the following two reasons to stay with master:

1. The change is perceived as part of a political movement that induces
   a counter-reaction in many people.
2. "master" is the already established name, and changing it might lead
   to confusion.

I can't relate to argument 1, and it's part of the political argument I
was going to ignore anyway. This leaves argument 2 which, in my opinion,
has been weakened severely by recent developments:

- GitHub and other hosts have already changed their default name for the
  main branch to "main".
- Many projects have changed the name of their `master` branch to `main`
  or something else.
- Git itself has added configuration that makes changing the name of the
  main branch for new repositories easy, and might or might not change
  the default name for the main branch in the future.

Given these development, I think argument 2 is basically void. Any
confusion this could produce is already here, so we're basically left
with the choice between an established and a new name. I'm choosing the
new one, for the reasons stated above.