Update description of Step/Dir
Originally, I planned Step/Dir as a low-level library over drivers with
STEP and DIR pins only, as the old version of the README indicated. This
plan didn't survive extended contact with reality. As it turns out, the
lines between drivers and controllers can be quite blurry, with many
drivers having controller-like features.

I've mulled over how to handle this, and came to the conclusion that
handling both drivers and controllers in the same library is the
solution that introduces the lowest friction overall.