Convert delay directly into timer ticks
Convert the delay return by the RampMaker profile directly into timer
ticks, instead of going through `embedded_time::Nanoseconds` as an
intermediate step.

Even though this saves one conversion (from the delay in nanoseconds to
timer ticks), it requires a new one (from the pulse length in
nanoseconds to timer ticks), leaving the same number of conversions in
the code. I still believes it is better, for two reasons:

1. The new conversion is redundant, as it is already done when
   performing the step pulse. I don't know how to re-use the result from
   that, but in principle, it should be possible.
2. This gives users the option to specify their velocities and
   accelerations in terms of timer ticks in the first place, completely
   saving the conversion from delay to timer ticks.